+++ crates/pallet-subspace/src/mock.rs
@@ -35,14 +35,15 @@ use sp_runtime::traits::{Block as BlockT, Header as _, IdentityLookup};
 use sp_runtime::Perbill;
 use std::sync::Once;
 use subspace_archiving::archiver::{ArchivedSegment, Archiver};
+use subspace_consensus_primitives::{
+    create_tag, create_tag_signature, derive_global_challenge, derive_local_challenge,
+    REWARD_SIGNING_CONTEXT,
+};
 use subspace_core_primitives::{
     ArchivedBlockProgress, LastArchivedBlock, LocalChallenge, Piece, Randomness, RootBlock, Salt,
     Sha256Hash, Solution, Tag, PIECE_SIZE,
 };
-use subspace_solving::{
-    create_tag, create_tag_signature, derive_global_challenge, derive_local_challenge,
-    SubspaceCodec, REWARD_SIGNING_CONTEXT,
-};
+use subspace_solving::SubspaceCodec;
 
 type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Test>;
 type Block = frame_system::mocking::MockBlock<Test>;
