+++ crates/subspace-consensus-primitives/Cargo.toml
@@ -13,6 +13,7 @@ readme = "README.md"
 
 [dependencies]
 codec = { package = "parity-scale-codec", version = "3.1.2", default-features = false }
+merlin = { version = "2.0.1", default-features = false }
 scale-info = { version = "2.1.1", default-features = false, features = ["derive"] }
 schnorrkel = { version = "0.9.1", default-features = false, features = ["u64_backend"] }
 sp-arithmetic = { version = "5.0.0", default-features = false, git = "https://github.com/subspace/substrate", rev = "b49d8ee54f4bd6e4b5b48f45d08f75705fad95b4" }
@@ -26,6 +27,7 @@ thiserror = { version = "1.0.31", optional = true }
 default = ["std"]
 std = [
     "codec/std",
+    "merlin/std",
     "scale-info/std",
     "schnorrkel/std",
     "sp-arithmetic/std",
