+++ crates/subspace-consensus-primitives/src/types.rs
@@ -1,8 +1,24 @@
+// Copyright (C) 2021 Subspace Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// 	http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 use codec::{Decode, Encode, MaxEncodedLen};
 use scale_info::TypeInfo;
 use subspace_core_primitives::{Randomness, Salt, Sha256Hash};
 
 /// Parameters for checking piece validity
+#[derive(Debug)]
 pub struct PieceCheckParams {
     /// Records root of segment to which piece belongs
     pub records_root: Sha256Hash,
@@ -17,6 +33,7 @@ pub struct PieceCheckParams {
 }
 
 /// Parameters for solution verification
+#[derive(Debug)]
 pub struct VerifySolutionParams<'a> {
     /// Global randomness
     pub global_randomness: &'a Randomness,
