+++ test/subspace-test-client/Cargo.toml
@@ -29,6 +29,7 @@ sp-core = { git = "https://github.com/subspace/substrate", rev = "b49d8ee54f4bd6
 sp-executor = { version = "0.1.0", path = "../../crates/sp-executor" }
 sp-runtime = { git = "https://github.com/subspace/substrate", rev = "b49d8ee54f4bd6e4b5b48f45d08f75705fad95b4" }
 subspace-archiving = { path = "../../crates/subspace-archiving" }
+subspace-consensus-primitives = { path = "../../crates/subspace-consensus-primitives" }
 subspace-core-primitives = { path = "../../crates/subspace-core-primitives" }
 subspace-runtime-primitives = { path = "../../crates/subspace-runtime-primitives" }
 subspace-service = { path = "../../crates/subspace-service" }
